requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
open-clip-torch>=2.24.0
faiss-cpu>=1.7.4
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
jupyter>=1.0.0
nbconvert>=7.0.0
requests>=2.31.0
pyyaml>=6.0
datasets>=2.14.0
optuna>=3.3.0
pytest>=7.4.0
